Reverse Digits
Reversing a number implies that the value of the unit of a given number will become the nth digit value of the number if it is n digits long, and the tens digit number will become (n-1)th place value number and so on. 

This program takes an integer as input, iteratively extracts the last digit (remainder), builds the reversed number by multiplying the current reversed number by 10 and adding the remainder, and finally removes the last digit from the original number. This process continues until the original number becomes zero.