Check Sum of Digit is Palindrome
Find the sum of digits of N and store it in a variable sum. Now check whether sum is palindrome or not.
checking palindrome:: A simple method for this problem is to first reverse digits of num, then compare the reverse of num with num. If both are same, then return true, else false.