Create Multiplication Table
vector<int> result;: This line declares a vector named result which will store integers (int).

for (int i = 1; i <= 10; ++i) { ... }: This is a for loop that iterates from 1 to 10, inclusive. It's used to generate the elements of the multiplication table.

result.push_back(N * i);: Inside the loop, N * i calculates the product of N (the given number) and i (the current iteration). result.push_back() adds this product to the end of the vector table.
return result;: Once the loop completes, the function returns the populated vector result, which contains the elements of the multiplication table for the given number N.