Calculate nPr
nPr represents n permutation r which is calculated as n!/(n-r)!. . 

factorial function:
This function calculates the factorial of a given integer n.
It's a recursive function. If n is less than or equal to 1, it returns 1 (as 0! and 1! are both 1). Otherwise, it multiplies n with the factorial of (n - 1).

nPr function:
This function calculates the permutation 𝑛𝑃r using the formula 
𝑛!/(𝑛−𝑟)!
​It calls the factorial function to calculate both n! and (n - r)!, then divides n! by (n - r)! to get the permutation value.