Find Simple Interest
The provided function simpleInterest calculates the simple interest using the formula:
Simple Interest=𝑃*𝑇*𝑅/100

However, there's a potential issue in the implementation. Since all the input parameters (P, R, T) are integers, the division operation may result in integer division, leading to loss of precision. To ensure that the division is performed in floating-point arithmetic, you should explicitly cast one of the operands to a floating-point type.

double(P * T * R) converts the product of P, T, and R to a double to ensure that the division operation is performed in floating-point arithmetic.